+++ packages/react-native/src/index.ts
@@ -1,4 +1,4 @@
-import { checkForUpdate } from "@hot-updater/utils";
+import { getUpdateInfo } from "@hot-updater/utils";
 import { NIL_UUID } from "./const";
 import { ensureBundles } from "./ensureBundles";
 import { init } from "./init";
@@ -29,7 +29,7 @@ export const HotUpdater = {
 
   ensureBundles,
   updateBundle,
-  checkForUpdate,
+  getUpdateInfo,
   /**
    * In production environment, this value will be replaced with a uuidv7.
    */